==> 02_medconf_practical.Rmd <==

# Mediation & Confounding

This section of the practical will take you through the initial analyses a social epidemiologist might take to determine if socioeconomic position is associated with HIV risk in young Tanzanian women. The goal of this practical is not to focus on the R or Stata code, but instead to understand conceptually what is happening and be able to interpret the output critically.

This initial session will consolidate your understanding of mediation and confounding. In this session we will seek to get to know the data, calculate some basic prevalences, describe the association between educational attainment and HIV, and adjust for a potential confounder.

There is no explicit paper associated with this practical, but some papers of interest are: 

In this practical, we will use data from the AIDS indicator survey carried out in Tanzania in 2007. The data includes only adolescent girls and young women aged 15-24 years of age who were residing in rural areas. The dataset contains one record for each of the 2762 young women included in the survey with information on their HIV serostatus.

Step 1 is to ensure that your R Notebook and your data are in the correct place. As long as the R Notebook and the dataset socepi_tz.dta (available from U:/Download/Teach/Social Epidemiology) are in the same place, the code below will work.

In R, not every tool we need for analysis is automatically included. We load a number of user-built libraries to make our lives easier. To run code in RStudio, you place your cursor on the line and press Ctrl-Enter, or if you want to run the whole chunk, mouse click the green triangle at the top right corner of the code chunk (in grey).


```{r medconf_setup, include = F}
# install.packages("readstata13")
# install.packages("epiDisplay")
# install.packages("magrittr")
# install.packages("tidyverse")

library(readstata13)
library(epiDisplay)
library(magrittr)
library(tidyverse)

#--- Turn off scientific notation and round everything to 3 decimal places
options(scipen = 999, digits = 3)

#--- Put all plots in black and white
theme_set(theme_bw())

```

Once you have loaded in the libraries, we are going to read in the data. You should see it appear in the 'Environment' pane in the top right hand corner of RStudio as a dataset called 'tz'. You can click on the dataset in the Environment pane and it will open in a new tab. Browse through the data now. 

```{r data_medconf}
#--- Read in data
tz <- read.dta("E:/LSHTM/Teaching/socepi/socepi_tz.dta", convert.factors = T)

```

## Descriptive Statistics

Let's do a quick summary of all of the variables. In R, we use the summary() command where Stata would use _summ_ or _tab_.

What is the age range of the population? What about the average age? Have a look at the below graph that describes the age range. 

The relevant commands for each of the variables are laid out. Take a look at their summaries. Make a rough table of your findings, perhaps including percentages for each variable. R has a built in calculator, so if you wanted to manually calculate a percentage, you can do that in the console pane on the lower left. Note any differences in variables that you think might be particularly important. Summarise verbally any patterns that you observe.

Make sure you intuit what each variable is measuring - think about whether it is a characteristic of the individual woman or of her household. 

```{r desc_medconf, warnings = F, message = F}
#--- Drop any factor levels that have no observations
tz <- droplevels(tz)

#--- Summaries of Age
summary(tz$age)
ggplot(tz, aes(x = as.factor(age))) +
  geom_bar()

#--- Summaries of other variables
summary(tz$serostat)
summary(tz$water)
summary(tz$toilet)
summary(tz$electricity)
summary(tz$radio)
summary(tz$tv)
summary(tz$fridge)
summary(tz$bike)
summary(tz$car)
summary(tz$floor)
summary(tz$wall)
summary(tz$roof)
summary(tz$educat)
summary(tz$married)
summary(tz$partners)

```

The below code recodes age into two different age groups, and the sexual partners variable into 0, 1, or 2+ sexual partners. 

From a measurement perspective, do you think this categorisation is useful or is there a better way to handle these variables? In other words, what are the implications of categorising these variables? Do you think this will affect your analysis? 


```{r recode_medconf}
#--- Recoding the age variable and specifying it is a factor
tz <- tz %>% mutate(age.group = ifelse(age < 20, "14-19", "20-24")) %>% 
             mutate(age.group = as.factor(age.group))

#--- Recoding the partner variable
tz <- tz %>% mutate(partners.cat = cut(partners,
                   breaks=c(-Inf, 0, 1, Inf), 
                   labels=c("0","1","2+")))
```

## HIV Prevalence

Now let's look at the prevalence of HIV in the sample. We can also examine HIV prevalence by age, education, and number of partners using either the _tabpct_ command or the _cc_ command. Use _tabpct_ to get row or column percentages and _cc_ to get an odds ratio and associated confidence interval. 

Do you notice any systematic differences in prevalence by these variables?


```{r summ_medconf}
#--- Stratified summaries of HIV prevalence by Age
tz %$% tabpct(serostat, age.group, percent = "col", graph = F)
tz %$% cc(serostat, age.group, graph = F)

#--- Are there trends in serostatus by age?
tz %>% ggplot(aes(x = as.factor(age), fill = serostat)) +
  geom_bar(position = "dodge")

#--- Stratified summaries of HIV prevalence by Education
tz %$% tabpct(educat, serostat, percent = "col", graph = F)
tz %$% cc(serostat, educat, graph = F) 

#--- Stratified summaries of HIV prevalence by Education
tz %$% tabpct(partners.cat, serostat, percent = "row", graph = F)
tz %$% cc(partners.cat, serostat, graph = F) 

```

## Age as Confounder

Consider whether or not age should be considered a confounding factor on the association between education and HIV.

What possible causal diagram represents the scenario where age is a confounder?

Compare the crude and adjusted ORs. Recalling that there is no formal statistical test for confounding, is there any _suggestion_ of confounding from the analyses below? 

```{r logreg_medconf}
#--- Logistic Regression
glm(serostat ~ educat + age.group, family = binomial, data = tz) %>% logistic.display() 
```

## Causal Diagrams

How might you extend your causal diagram to include other potential confounders (or mediators) from this dataset? For example, what happens when we also investigate the potential confounding effect of marriage? 

On your causal diagram, which direction does the arrow go between education and marriage? How did you decide this?

```{r logreg_adj_medconf}
#--- Logistic Regression
glm(serostat ~ educat + age.group + married, family = binomial, data = tz) %>% logistic.display() 
```

Using your own knowledge and that of your colleagues, or by doing a brief search of the literature, list some potential mechanisms by which education could affect HIV risk in young women in Tanzania. A mechanism is some route by which which change in an outcome is created; the process by which causal effects take place. For example, one mechanism could be education impacting HIV risk because HIV prevention programmes are delivered in schools.

What might the causal diagram look like for each of your identified mechanisms? Are any of your identified mechanisms of action specific to a particular context? Are any of these mechanisms testable from your data? 

In the example above, our causal diagram would be amended to include a mediating variable between education and HIV risk. This mechanism would only be relevant in the context of girls who have been to school. We cannot test for this mechanism in the data as we do not have knowledge of whether HIV prevention programmes were implemented or not in each school. Note that we could collect this if we were designing an intervention or a survey - it is key in practical social epidemiology to consider possible mechanisms of action in the design phase. It is often insufficient to ask simply _whether_ there is an effect of a particular construct - we wish to know _how_ and _why_ it exists. 

## Save New Dataset

We have made some changes to the dataset that we may wish to save to use later.

```{r write_medconf}
write.csv(tz, "tz2.csv", row.names = F)
```
